+++ regression/reference/names2/wrapNames.h
@@ -28,7 +28,7 @@ extern "C" {
 // splicer begin C_declarations
 // splicer end C_declarations
 
-void NAM_afunction(void);
+void NAM_a_function(void);
 
 #ifdef __cplusplus
 }