+++ regression/reference/struct-c/wrapstruct.h
@@ -29,15 +29,15 @@ int STR_pass_struct2_bufferify(const Cstruct1 * s1, char *outbuf,
 Cstruct1 * STR_return_struct_ptr2_bufferify(int i, double d,
     char *outbuf, int SHT_outbuf_len);
 
-STR_Cstruct_as_class * STR_create__cstruct_as_class(
+STR_Cstruct_as_class * STR_create_cstruct_as_class(
     STR_Cstruct_as_class * SHC_rv);
 
-STR_Cstruct_as_class * STR_create__cstruct_as_class_args(int x, int y,
+STR_Cstruct_as_class * STR_create_cstruct_as_class_args(int x, int y,
     STR_Cstruct_as_class * SHC_rv);
 
 int STR_cstruct_as_class_sum(STR_Cstruct_as_class * point);
 
-STR_Cstruct_as_subclass * STR_create__cstruct_as_subclass_args(int x,
+STR_Cstruct_as_subclass * STR_create_cstruct_as_subclass_args(int x,
     int y, int z, STR_Cstruct_as_subclass * SHC_rv);
 
 const double * STR_cstruct_ptr_get_const_dvalue_bufferify(